requirements.txt
amqp==2.6.0
billiard==3.6.3.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
charset-normalizer==3.3.2
cmsplugin-filer==1.1.3
cssselect2==0.7.0
Django==2.2.14
django-appconf==1.0.4
django-classy-tags==1.0.0
django-cms==3.7.4
django-cors-headers==2.5.3
django-filer==1.7.1
django-formtools==2.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.1.2
django-sekizai==1.1.0
django-treebeard==4.3.1
djangocms-admin-style==1.5.0
djangocms-attributes-field==1.2.0
djangocms-column==1.10.0
djangocms-googlemap==1.4.0
djangocms-link==2.6.1
djangocms-snippet==2.3.0
djangocms-style==2.3.0
djangocms-text-ckeditor==3.9.1
djangocms-video==2.3.0
djangorestframework==3.11.0
earthengine-api==0.1.229
easy-thumbnails==2.7
future==0.18.2
google-api-core==1.22.0
google-api-python-client==1.10.0
google-auth==1.20.0
google-auth-httplib2==0.0.4
google-cloud-core==1.3.0
google-cloud-storage==1.30.0
google-crc32c==0.1.0
google-resumable-media==0.7.0
googleapis-common-protos==1.52.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.18.1
httplib2shim==0.0.3
idna==2.10
importlib-metadata==1.7.0
kombu==4.6.11
lxml==5.2.2
Markdown==2.6.11
numpy==2.0.1
oauth2client==4.1.3
packaging==24.0
pilkit==2.0
Pillow==9.0.0
pinax-blog==8.0.1
pinax-images==4.0.1
proto-plus==1.23.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.6.1
pyparsing==3.1.2
pytz==2020.1
reportlab==4.2.0
requests==2.24.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.10
vine==1.3.0
webencodings==0.5.1
wincertstore==0.2
zipp==3.1.0
